+++ site/dat/docs/Changelog.md
@@ -6,6 +6,7 @@
  - do not create hostaliases file when aliases are not specified
  - add `force-parent-sample` option to JMeter executor
  - add `compile-target-java` option for Selenium
+ - add dynamic buffer scaling for results reader
 
 ## 1.3.3 <sup>24 mar 2016</sup>
  - add new `hostaliases` setting for all executors
@@ -14,7 +15,6 @@
 ## 1.3.2 <sup>23 mar 2016</sup>
  - fix lowercase hostname for JMeter HTTP request
  - fix binary varname crash for JMeter HTTP request 
- - add dynamic buffer scaling for results reader
 
 ## 1.3.1 <sup>22 mar 2016</sup>
  - fix JMeter crash when `data-sources` value is not a list